Create README.md
DECRIPTION

The simplest syslog daemon and client
It is working with REUSEADDR socket option. 
So service is set of separated processes.
Each syslog process is listening in his own 
address space and writing to common log file /var/log/syslogs.log.

The format of client syslogc is not RFC compatible, 
it is very simple for writing.
It can be changed for needs of compatibility in the future.
 
COMPILE

make

USING

1) Start single daemon:

sudo make start

or

./syslogs > /var/log/syslogs.log

2) test client

./syslogc "Some Test Record"

3) Start multyple instances of syslog:

./run.sh

4) Run with foring supervisor:

./supervisor

5) Run syslog with internal supervising:

./syslogs3

This script ensures what working 8 syslogs and start if needed.

FAQ:

Q1) Whatis the difference between syslogs and syslogs2
A1) The syslogs2 is more complicated.
    It :
    *) Reads and parse preambule of message, 
       open log's file , related with it;
    *) Remove preambule, write message to this file

    So all syslog messages will be distributed inside a set of syslogs files.
    Each kind of preambule <N> will be inside it's own file.
  
    While more simple syslogs writes incoming messages as is, without any
    transformation.

Q2) Can it be faster ?
A2) Shure. 
    The syslogs shuld be faster than syslogs2, it do not any transforms or parsing 
    for message.
    The syslogs2 is reopen log file for each incoming message, and parse it for
    computation of log file name, but syslogs2 can works in parallel mode and
    it should compensate long log writing.

Q3) Where it can be used ?
A3) For education, for systems where needs robast and fault tolerance events
    logging and handling.